reimplement `Score` using a single i32 for performance
This commit contains both the old `Score` implementation (module `score`), based on an enum with 4 variants, as well as the new `Score` implementation (module `score2`), which uses just a single i32. It achieves this by using different ranges of values to represent centipawn vs. mate-in-N scores. Performance is significantly better as most operations are now simple comparisons rather than enum match statements with loads of arms.